Fix MinGW printf format errors for size_t
 MinGW-w64 on certain platforms does not correctly handle the %zu format specifier for size_t, leading to build errors when warnings are treated as errors.

This change replaces %zu with %llu and casts the corresponding arguments to (unsigned long long) for better cross-platform compatibility.

Fixes #3039

+++ src/user/user_vfs.cc
@@ -130,8 +130,8 @@ VFS::VFS(mjVFS* vfs) : self_(vfs) {
 VFS::~VFS() {
   if (!open_resources_.empty()) {
     mju_warning(
-        "VFS destroyed with %zu open resources. Resources will be invalidated.",
-        open_resources_.size());
+        "VFS destroyed with %llu open resources. Resources will be invalidated.",
+        (unsigned long long)open_resources_.size());
   }
   for (auto& [ptr, res] : open_resources_) {
     if (res->provider->close) {
